+++ BaseUriTest.php
@@ -360,6 +360,10 @@ final class BaseUriTest extends TestCase
                 'uri' => Uri::new('https://0:443/'),
                 'expectedOrigin' => 'https://0.0.0.0',
             ],
+            'compressed ipv6' => [
+                'uri' => 'https://[1050:0000:0000:0000:0005:0000:300c:326b]:443/',
+                'expectedOrigin' => 'https://[1050::5:0:300c:326b]',
+            ],
         ];
     }